requirements.txt
jpype1==1.5.0
aspose-cells==23.7.0
